refactor(scoring): extract scoring helpers into own module
